# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A meow API server built with Go, inspired by meow",
        "title": "meow meow API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "meow API Support",
            "url": "https://github.com/your-username/meow",
            "email": "support@meow.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/session/{sessionId}/community/link": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Link a group to a community",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community"
                ],
                "summary": "Link group to community",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Link group request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.LinkGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunityResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunityResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunityResponse"
                        }
                    }
                }
            }
        },
        "/session/{sessionId}/community/participants": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all participants of linked groups in a community",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community"
                ],
                "summary": "Get community participants",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Get participants request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetLinkedGroupsParticipantsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunityParticipantsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunityResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunityResponse"
                        }
                    }
                }
            }
        },
        "/session/{sessionId}/community/subgroups": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get all subgroups of a community",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community"
                ],
                "summary": "Get community subgroups",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Get subgroups request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.GetSubGroupsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunitySubGroupsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunityResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunityResponse"
                        }
                    }
                }
            }
        },
        "/session/{sessionId}/community/unlink": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Unlink a group from a community",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Community"
                ],
                "summary": "Unlink group from community",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Session ID",
                        "name": "sessionId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Unlink group request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UnlinkGroupRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunityResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunityResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/dto.CommunityResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CommunityData": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "link_group"
                },
                "community_jid": {
                    "type": "string",
                    "example": "120363025246125486@g.us"
                },
                "group_jid": {
                    "type": "string",
                    "example": "120363025246125487@g.us"
                },
                "session_id": {
                    "type": "string",
                    "example": "default"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                }
            }
        },
        "dto.CommunityErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INVALID_COMMUNITY_JID"
                },
                "details": {
                    "type": "string",
                    "example": "Community JID must be in format: number@g.us"
                },
                "message": {
                    "type": "string",
                    "example": "Invalid community JID format"
                }
            }
        },
        "dto.CommunityParticipantsData": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "get_participants"
                },
                "community_jid": {
                    "type": "string",
                    "example": "120363025246125486@g.us"
                },
                "participants": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"5511999999999\"",
                        " \"5511888888888\"]"
                    ]
                },
                "session_id": {
                    "type": "string",
                    "example": "default"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "total": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "dto.CommunityParticipantsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/dto.CommunityParticipantsData"
                },
                "error": {
                    "$ref": "#/definitions/dto.CommunityErrorResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.CommunityResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/dto.CommunityData"
                },
                "error": {
                    "$ref": "#/definitions/dto.CommunityErrorResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "dto.CommunitySubGroupsData": {
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "get_subgroups"
                },
                "community_jid": {
                    "type": "string",
                    "example": "120363025246125486@g.us"
                },
                "session_id": {
                    "type": "string",
                    "example": "default"
                },
                "status": {
                    "type": "string",
                    "example": "success"
                },
                "subgroups": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"120363025246125487@g.us\"",
                        " \"120363025246125488@g.us\"]"
                    ]
                },
                "timestamp": {
                    "type": "string",
                    "example": "2023-01-01T00:00:00Z"
                },
                "total": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "dto.CommunitySubGroupsResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "$ref": "#/definitions/dto.CommunitySubGroupsData"
                },
                "error": {
                    "$ref": "#/definitions/dto.CommunityErrorResponse"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.GetLinkedGroupsParticipantsRequest": {
            "type": "object",
            "required": [
                "community_jid"
            ],
            "properties": {
                "community_jid": {
                    "type": "string",
                    "example": "120363025246125486@g.us"
                }
            }
        },
...
        "dto.GetSubGroupsRequest": {
            "type": "object",
            "required": [
                "community_jid"
            ],
            "properties": {
                "community_jid": {
                    "type": "string",
                    "example": "120363025246125486@g.us"
                }
            }
        },
...
        "dto.LinkGroupRequest": {
            "type": "object",
            "required": [
                "community_jid",
                "group_jid"
            ],
            "properties": {
                "community_jid": {
                    "type": "string",
                    "example": "120363025246125486@g.us"
                },
                "group_jid": {
                    "type": "string",
                    "example": "120363025246125487@g.us"
                }
            }
        },
...
        "dto.UnlinkGroupRequest": {
            "type": "object",
            "required": [
                "community_jid",
                "group_jid"
            ],
            "properties": {
                "community_jid": {
                    "type": "string",
                    "example": "120363025246125486@g.us"
                },
                "group_jid": {
                    "type": "string",
                    "example": "120363025246125487@g.us"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "API Key authentication. Simply provide your API key directly: \"YOUR_API_KEY\". The system automatically detects if it's a Global API Key (can access all sessions and session management) or a Session-specific API Key (can only access the specific session it belongs to).",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}